fix: warning解消

+++ DirectX11_Base/Source/System/ryuu_lib/IME_Controller/IME.cpp
@@ -134,7 +134,7 @@ void IME::SetCompositionWindowPos(const SF2Position& In_f2Pos) noexcept
 	}
 }
 
-void IME::SetCompositionFont(int In_nFontSizeHeight, int In_nFontSizeWidth, BYTE In_lfWeight,
+void IME::SetCompositionFont(int In_nFontSizeHeight, int In_nFontSizeWidth, LONG In_lfWeight,
 	BYTE In_lfItalic, BYTE In_lfUnderline, BYTE In_lfStrikeOut, BYTE In_lfCharSet)
 {
 	LPLOGFONTA lpLogFont = new LOGFONTA(); // LOGFONTAÇ≈èâä˙âª
